refactor(func): sharing array func
Let the array func and aggregate func share the implementation if possible.
